backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
baidusearch==1.0.3
beartype==0.22.5
beautifulsoup4==4.14.2
blinker==1.9.0
blockbuster==1.5.25
boto3==1.40.76
botocore==1.40.76
browser-use==0.5.11
bubus==1.5.6
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
copilotkit
coverage==7.12.0
cryptography==44.0.3
cyclopts==4.2.4
Cython==3.2.1
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
duckduckgo_search==8.1.1
e2b==1.5.0
e2b-code-interpreter==1.5.0
e2b-desktop==1.7.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.115.14
fastmcp==2.13.1
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
forbiddenfruit==0.1.4
frozenlist==1.8.0
fsspec==2025.10.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
google-genai==1.29.0
google_search_results==2.4.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
groq==0.35.0
grpcio
grpcio-tools
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
json5==0.12.1
json_repair==0.54
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jsonschema_rs==0.29.1
keyring==25.7.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-community
langchain-core>=0.1.0
langchain-mcp-adapters==0.1.13
langchain-openai==0.3.29
langchain-text-splitters==0.3.11
langgraph>=0.0.20,<0.3.0
langgraph-api==0.0.48
langgraph-checkpoint==2.1.2
langgraph-cli==0.1.71
langgraph-prebuilt<1.0.0
langgraph-runtime-inmem
langgraph-sdk<0.2.0
langsmith==0.3.45
litellm==1.52.8
lxml==6.0.2
magika==0.6.3
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown[all]==0.1.3
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.21.2
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.5
oauthlib==3.3.1
ollama==0.6.1
openai==1.99.2
openapi-pydantic==0.5.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
partialjson==0.0.8
pathable==0.4.4
pathvalidate==3.3.1
pillow==11.3.0
platformdirs==4.5.0
playwright==1.56.0
pluggy==1.6.0
portalocker==2.10.1
postgrest==2.24.0
posthog==7.0.1
primp==0.15.0
propcache==0.4.1
proto-plus==1.26.1
protobuf
psutil==7.1.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.3.0
pyperclip==1.11.0
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-a2a==0.5.9
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
realtime==2.24.0
redis==7.0.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
screeninfo==0.8.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sse-starlette==2.1.3
starlette==0.46.2
storage3==2.24.0
StrEnum==0.4.15
structlog==25.5.0
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.14.0
tavily-python==0.7.13
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.21.4
toml==0.10.2
tqdm==4.67.1
transformers==4.51.3
truststore==0.10.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid7==0.1.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0
